requirements.txt
asgiref==3.9.1
cffi==1.17.1
cryptography==45.0.5
Django==4.2.23
django-appconf==1.1.0
django-encrypted-model-fields==0.6.5
django-unfold==0.63.0
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==11.3.0
pycparser==2.22
sqlparse==0.5.3
typing_extensions==4.14.1
